+++ src/main/java/com/mohinhphanlop/projectthree/Repositories/ThanhVienRepository.java
@@ -17,5 +17,5 @@ public interface ThanhVienRepository extends CrudRepository<ThanhVien, Integer> 
     List<ThanhVien> GetListThanhVienTheoNam(String year);
 
     @Query("SELECT TV.maTV FROM ThanhVien TV WHERE TV.maTV NOT IN (SELECT TT.thanhvien.maTV FROM ThongTinSD TT WHERE tGTra IS NULL AND tGVao IS NULL) AND TV.maTV NOT IN (SELECT XL.thanhvien.maTV FROM XuLy XL WHERE trangThaiXL = 0) AND SUBSTRING(CAST(TV.maTV AS string),3,2) = :lastSchoolYear")
-    List<Long> GetIDHopLe(String lastSchoolYear);
+    List<Integer> GetIDHopLe(String lastSchoolYear);
 }
